django_server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
django==4.1.3
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.13.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
mysqlclient==2.1.1
    # via -r requirements.in
pytz==2022.7.1
    # via djangorestframework
sqlparse==0.4.3
    # via django
